Move batched loader into batch.go
